refactor: cleanup environments pages
