+++ frontend/messages/en.json
@@ -1261,7 +1261,6 @@
 	"environments_created_success": "Environment created successfully",
 	"environments_delete_message": "Are you sure you want to delete environment {name}?",
 	"environments_delete_failed": "Failed to delete environment {name}",
-	"environments_overview_title": "Environment Overview",
 	"environments_environment_id_label": "Environment ID",
 	"environments_name_placeholder": "My Lab Server",
 	"environments_name_too_long": "Name too long (max 25 characters)",
@@ -1271,7 +1270,6 @@
 	"environments_create_failed": "Failed to create environment",
 	"environments_unsaved_changes": "Unsaved changes",
 	"environments_all_changes_saved": "All changes saved",
-	"environments_basic_info_description": "Basic environment information and settings",
 	"environments_enable_disable_description": "Enable or disable this environment",
 	"environments_local_badge": "Local",
 	"environments_edge_http_label": "HTTP",
