--- frontend/src/routes/(app)/environments/[id]/components/DetailsTab.svelte
@@ -1,345 +1,0 @@
-<script lang="ts">
-	import { format } from 'date-fns';
-	import * as Card from '$lib/components/ui/card/index.js';
-	import Label from '$lib/components/ui/label/label.svelte';
-	import { Input } from '$lib/components/ui/input/index.js';
-	import { Badge } from '$lib/components/ui/badge/index.js';
-	import { ArcaneButton } from '$lib/components/arcane-button/index.js';
-	import * as ArcaneTooltip from '$lib/components/arcane-tooltip';
-	import StatusBadge from '$lib/components/badges/status-badge.svelte';
-	import { Spinner } from '$lib/components/ui/spinner/index.js';
-	import { m } from '$lib/paraglide/messages';
-	import { EnvironmentsIcon, TestIcon } from '$lib/icons';
-	import type { DetailsTabProps } from './tab-props';
-
-	let {
-		environment,
-		formInputs,
-		currentStatus,
-		isLoadingVersion,
-		remoteVersion,
-		versionInformation,
-		isTestingConnection,
-		testConnection
-	}: DetailsTabProps = $props();
-
-	let transportBadge = $derived.by((): { text: string; variant: 'blue' | 'purple' | 'gray' } => {
-		if (!environment.isEdge) {
-			return { text: 'HTTP', variant: 'gray' };
-		}
-
-		// Prefer the live tunnel transport; fall back to the last one used so
-		// disconnected or poll-only agents still show what they connect with.
-		const transport = (environment.connected ? environment.edgeTransport : undefined) ?? environment.lastEdgeTransport;
-		if (!transport) {
-			return { text: 'Edge', variant: 'gray' };
-		}
-
-		if (transport === 'websocket') {
-			return { text: 'WebSocket', variant: 'purple' };
-		}
-
-		return { text: 'gRPC', variant: 'blue' };
-	});
-
-	let controlPlaneBadge = $derived.by((): { text: string; variant: 'blue' | 'green' | 'gray' } | null => {
-		if (!environment.isEdge || !environment.lastPollAt) {
-			return null;
-		}
-
-		if (environment.connected) {
-			return { text: m.environments_edge_polling_active(), variant: 'green' };
-		}
-
-		if (currentStatus === 'standby') {
-			return { text: m.environments_edge_polling_standby(), variant: 'blue' };
-		}
-
-		return { text: m.environments_edge_polling_inactive(), variant: 'gray' };
-	});
-
-	let localDisplayVersion = $derived(
-		versionInformation?.displayVersion || versionInformation?.currentTag || versionInformation?.currentVersion || 'Unknown'
-	);
-
-	let remoteDisplayVersion = $derived(
-		remoteVersion?.displayVersion || remoteVersion?.currentTag || remoteVersion?.currentVersion || ''
-	);
-
-	let statusBadge = $derived.by((): { text: string; variant: 'green' | 'blue' | 'amber' | 'red' } => {
-		switch (currentStatus) {
-			case 'online':
-				return { text: m.common_online(), variant: 'green' };
-			case 'standby':
-				return { text: m.common_standby(), variant: 'blue' };
-			case 'pending':
-				return { text: m.common_pending(), variant: 'amber' };
-			case 'error':
-				return { text: m.common_error(), variant: 'red' };
-			default:
-				return { text: m.common_offline(), variant: 'red' };
-		}
-	});
-
-	let mtlsCertificateBadge = $derived.by((): { text: string; variant: 'green' | 'amber' | 'red' } | null => {
-		const cert = environment.edgeMTLSCertificate;
-		if (!cert) return null;
-		if (cert.expired) {
-			return { text: m.environments_edge_mtls_certificate_status_expired(), variant: 'red' };
-		}
-		if (cert.expiringSoon) {
-			return { text: m.environments_edge_mtls_certificate_status_expiring_soon(), variant: 'amber' };
-		}
-		return { text: m.environments_edge_mtls_certificate_status_valid(), variant: 'green' };
-	});
-
-	let tunnelBadge = $derived.by((): { text: string; variant: 'green' | 'blue' | 'gray' | 'amber' | 'red' } => {
-		if (!environment.isEdge) {
-			return statusBadge;
-		}
-		if (environment.connected) {
-			return { text: m.environments_edge_tunnel_transmitting(), variant: 'green' };
-		}
-		if (currentStatus === 'standby') {
-			return { text: m.environments_edge_tunnel_dormant(), variant: 'gray' };
-		}
-		if (currentStatus === 'pending') {
-			return { text: m.environments_edge_tunnel_negotiating(), variant: 'amber' };
-		}
-		return { text: m.environments_edge_tunnel_disconnected(), variant: 'red' };
-	});
-
-	let tunnelTypeBadge = $derived.by((): { text: string; variant: 'blue' | 'purple' | 'gray' } | null => {
-		if (!environment.isEdge || !environment.lastPollAt) {
-			return null;
-		}
-
-		if (environment.edgeTransport === 'websocket') {
-			return { text: 'WebSocket', variant: 'purple' };
-		}
-
-		if (environment.edgeTransport === 'grpc') {
-			return { text: 'gRPC', variant: 'blue' };
-		}
-
-		return { text: m.environments_edge_tunnel_type_inactive(), variant: 'gray' };
-	});
-
-	function formatDateTime(value?: string): string {
-		if (!value) return m.common_never();
-
-		const date = new Date(value);
-		if (Number.isNaN(date.getTime())) {
-			return m.common_unknown();
-		}
-
-		return format(date, 'PP p');
-	}
-</script>
-
-<Card.Root class="flex flex-col">
-	<Card.Header icon={EnvironmentsIcon}>
-		<div class="flex flex-col space-y-1.5">
-			<Card.Title>
-				<h2>{m.environments_overview_title()}</h2>
-			</Card.Title>
-			<Card.Description>{m.environments_basic_info_description()}</Card.Description>
-		</div>
-	</Card.Header>
-	<Card.Content class="space-y-4 p-4">
-		<div>
-			<Label for="env-name" class="text-sm font-medium">{m.common_name()}</Label>
-			<Input
-				id="env-name"
-				type="text"
-				bind:value={$formInputs.name.value}
-				class="mt-1.5 w-full {$formInputs.name.error ? 'border-destructive' : ''}"
-				placeholder={m.environments_name_placeholder()}
-			/>
-			{#if $formInputs.name.error}
-				<p class="text-destructive mt-1 text-[0.8rem] font-medium">{$formInputs.name.error}</p>
-			{/if}
-		</div>
-
-		<div>
-			<Label for="api-url" class="text-sm font-medium">{m.environments_api_url()}</Label>
-			<div class="mt-1.5 flex items-center gap-2">
-				{#if environment.id === '0'}
-					<ArcaneTooltip.Root>
-						<ArcaneTooltip.Trigger class="w-full">
-							<Input
-								id="api-url"
-								type="url"
-								bind:value={$formInputs.apiUrl.value}
-								class="w-full font-mono"
-								placeholder={m.environments_api_url_placeholder()}
-								disabled={true}
-								required
-							/>
-						</ArcaneTooltip.Trigger>
-						<ArcaneTooltip.Content>
-							<p>{m.environments_local_setting_disabled()}</p>
-						</ArcaneTooltip.Content>
-					</ArcaneTooltip.Root>
-				{:else}
-					<Input
-						id="api-url"
-						type="url"
-						bind:value={$formInputs.apiUrl.value}
-						class="w-full font-mono"
-						placeholder={m.environments_api_url_placeholder()}
-						required
-					/>
-				{/if}
-				<ArcaneButton
-					action="base"
-					onclick={testConnection}
-					disabled={isTestingConnection}
-					loading={isTestingConnection}
-					icon={TestIcon}
-					customLabel={m.environments_test_connection()}
-					loadingLabel={m.environments_testing_connection()}
-					class="shrink-0"
-				/>
-			</div>
-			<p class="text-muted-foreground mt-1.5 text-xs">{m.environments_api_url_help()}</p>
-		</div>
-
-		<div class="grid grid-cols-2 gap-4 rounded-lg border p-4">
-			<div>
-				<Label class="text-muted-foreground text-xs font-medium">{m.environments_environment_id_label()}</Label>
-				<div class="mt-1 font-mono text-sm">{environment.id}</div>
-			</div>
-			<div>
-				<Label class="text-muted-foreground text-xs font-medium">{m.common_status()}</Label>
-				<div class="mt-1">
-					<StatusBadge text={statusBadge.text} variant={statusBadge.variant} />
-				</div>
-			</div>
-			<div>
-				<Label class="text-muted-foreground text-xs font-medium">{m.common_type()}</Label>
-				<div class="mt-1">
-					<StatusBadge text={transportBadge.text} variant={transportBadge.variant} />
-				</div>
-			</div>
-			{#if environment.isEdge}
-				{#if controlPlaneBadge}
-					<div>
-						<Label class="text-muted-foreground text-xs font-medium">{m.environments_edge_control_plane_label()}</Label>
-						<div class="mt-1">
-							<StatusBadge text={controlPlaneBadge.text} variant={controlPlaneBadge.variant} />
-						</div>
-					</div>
-					<div>
-						<Label class="text-muted-foreground text-xs font-medium">{m.environments_edge_last_poll_label()}</Label>
-						<div class="mt-1 font-mono text-sm">{formatDateTime(environment.lastPollAt)}</div>
-					</div>
-				{/if}
-				<div>
-					<Label class="text-muted-foreground text-xs font-medium">{m.environments_edge_live_tunnel_label()}</Label>
-					<div class="mt-1">
-						<StatusBadge text={tunnelBadge.text} variant={tunnelBadge.variant} />
-					</div>
-				</div>
-				{#if tunnelTypeBadge}
-					<div>
-						<Label class="text-muted-foreground text-xs font-medium">{m.environments_edge_tunnel_type_label()}</Label>
-						<div class="mt-1">
-							<StatusBadge text={tunnelTypeBadge.text} variant={tunnelTypeBadge.variant} />
-						</div>
-					</div>
-				{/if}
-				<div>
-					<Label class="text-muted-foreground text-xs font-medium">{m.environments_edge_connected_since_label()}</Label>
-					<div class="mt-1 font-mono text-sm">{formatDateTime(environment.connectedAt)}</div>
-				</div>
-				<div>
-					<Label class="text-muted-foreground text-xs font-medium">{m.environments_edge_last_heartbeat_label()}</Label>
-					<div class="mt-1 font-mono text-sm">{formatDateTime(environment.lastHeartbeat)}</div>
-				</div>
-			{/if}
-			<div class="col-span-2 border-t pt-4">
-				<Label class="text-muted-foreground text-xs font-medium">{m.version_info_version()}</Label>
-				<div class="mt-1 flex items-center gap-2">
-					{#if environment.id === '0'}
-						<span class="font-mono text-sm">{localDisplayVersion}</span>
-						{#if versionInformation?.updateAvailable}
-							<Badge variant="secondary" class="bg-amber-500/10 text-amber-600 hover:bg-amber-500/20 dark:text-amber-400">
-								{m.sidebar_update_available()}: {versionInformation.newestVersion}
-							</Badge>
-						{/if}
-					{:else if isLoadingVersion}
-						<Spinner />
-						<span class="text-muted-foreground text-sm">{m.common_action_checking()}</span>
-					{:else if remoteVersion}
-						<span class="font-mono text-sm">{remoteDisplayVersion}</span>
-						{#if remoteVersion.updateAvailable}
-							<Badge variant="secondary" class="bg-amber-500/10 text-amber-600 hover:bg-amber-500/20 dark:text-amber-400">
-								{m.sidebar_update_available()}: {remoteVersion.newestVersion}
-							</Badge>
-							{#if remoteVersion.releaseUrl}
-								<a
-									href={remoteVersion.releaseUrl}
-									target="_blank"
-									rel="noopener noreferrer"
-									class="text-xs text-blue-500 hover:underline"
-								>
-									{m.version_info_view_release()}
-								</a>
-							{/if}
-						{/if}
-					{:else if currentStatus === 'online' || currentStatus === 'standby'}
-						<span class="text-muted-foreground text-sm">{m.environments_version_unavailable()}</span>
-					{:else}
-						<span class="text-muted-foreground text-sm">{m.common_offline()}</span>
-					{/if}
-				</div>
-			</div>
-		</div>
-
-		{#if mtlsCertificateBadge && environment.edgeMTLSCertificate}
-			<div class="space-y-3 rounded-lg border p-4">
-				<div class="space-y-1">
-					<h3 class="text-sm font-medium">{m.environments_agent_mtls_section_title()}</h3>
-					<p class="text-muted-foreground text-xs">{m.environments_agent_mtls_description()}</p>
-				</div>
-
-				<div class="grid grid-cols-1 gap-4 sm:grid-cols-2">
-					<div>
-						<Label class="text-muted-foreground text-xs font-medium">
-							{m.environments_edge_mtls_certificate_status_label()}
-						</Label>
-						<div class="mt-1">
-							<StatusBadge text={mtlsCertificateBadge.text} variant={mtlsCertificateBadge.variant} />
-						</div>
-					</div>
-					<div>
-						<Label class="text-muted-foreground text-xs font-medium">
-							{m.environments_edge_mtls_certificate_expires_label()}
-						</Label>
-						<div class="mt-1 space-y-1">
-							<div class="font-mono text-sm">
-								{environment.edgeMTLSCertificate.expiresAt ? formatDateTime(environment.edgeMTLSCertificate.expiresAt) : '—'}
-							</div>
-							{#if environment.edgeMTLSCertificate.daysRemaining !== undefined}
-								<div class="text-muted-foreground text-xs">
-									{m.environments_edge_mtls_certificate_days_remaining({
-										count: environment.edgeMTLSCertificate.daysRemaining
-									})}
-								</div>
-							{/if}
-						</div>
-					</div>
-					{#if environment.edgeMTLSCertificate.commonName}
-						<div class="sm:col-span-2">
-							<Label class="text-muted-foreground text-xs font-medium">
-								{m.environments_edge_mtls_certificate_common_name_label()}
-							</Label>
-							<div class="mt-1 font-mono text-sm">{environment.edgeMTLSCertificate.commonName}</div>
-						</div>
-					{/if}
-				</div>
-			</div>
-		{/if}
-	</Card.Content>
-</Card.Root>
