+++ frontend/src/routes/(app)/environments/[id]/components/tab-props.ts
@@ -1,24 +1,17 @@
 import type { Writable } from 'svelte/store';
 import type { FormInputs } from '$lib/utils/settings';
 import type { Environment, EnvironmentStatus } from '$lib/types/environment';
-import type { AppVersionInformation } from '$lib/types/settings';
 import type { EnvironmentFormValues } from './environment-form-schema';
 
 export type EnvironmentFormInputs = Writable<FormInputs<EnvironmentFormValues>>;
 
-export interface DetailsTabProps {
-	environment: Environment;
-	formInputs: EnvironmentFormInputs;
-	currentStatus: EnvironmentStatus;
-	isLoadingVersion: boolean;
-	remoteVersion: AppVersionInformation | null;
-	versionInformation: AppVersionInformation | null | undefined;
-	isTestingConnection: boolean;
-	testConnection: () => void | Promise<void>;
-}
-
 export interface GeneralTabProps {
 	formInputs: EnvironmentFormInputs;
+	environment: Environment;
+	currentStatus: EnvironmentStatus;
+	isTestingConnection: boolean;
+	testConnection: () => void | Promise<void>;
+	settingsAvailable: boolean;
 }
 
 export interface DockerTabProps {
